Disable Webpack caching
Webpack runs in development mode because it makes the output easier to
read for humans. This may change in the future but for now it's really
useful. However, in development mode Webpack caches things in the same
process. This means that Webpack did not always recompile snippets.

Since the single snippet test and the multiple snippet test shared a
snippet that was identical, Webpack had cached the work and didn't do it
twice. This caused the result file to be unreadable from disk however,
which caused our program to crash, resulting in difficult to debug
testing behaviour.