Enable support for dependencies in snippets
This commit allows a user to specify a dependency in a snippet using an
annotation. This dependency is then loaded by the bsb build process.

To accomodate this a test package is made and linked. In a future commit
this package should be included in the repository but that was a bit too
much work for now.

This step also introduces the possibility to change the bsb
configuration per snippet. The current dependency test still causes a
webpack warning and an error when executing because the default embed
snippet tries to load a React component but the test snippet just prints
"Hello World" to the console and doesn't have a react component.

In a future commit this "prelude" will need to be changeable.