Move embed snippet out of code
This now allows the embed snippet to be changed so that more fancy
options than a simple script tag can be added. Additionally this paves
the way of having other JavaScript snippets be used as preludes.

For example, currently it's assumed that the snippet always exports a
React component, but that's not necessarily true. This causes errors
with the import in the embed code.