Add support for prelude switching
The prelude is the code snippet that is compiled by webpack and includes
the code snippet that is defined in markdown. It still defaults to a
React component but it can now be changed with the `@prelude`
annotation.

The prelude of `none` has been added which simply imports a script
without providing any other code. This is useful if you want to
demonstrate snippets that work with the DOM or only produce textual
output.