Remove files from package.json and update npmignore
The files directive causes yarn to ignore .npmignore altogether. This
causes quite a bit of test data to be published with the package which
is annoying.

Instead we set up .npmignore to blacklist all of this for us. This is
slightly less secure but works better until the Yarn issue is fixed.

https://github.com/yarnpkg/yarn/issues/685