Add script-execution in development mode
Gatsby sets the innerHTML for markdown pages during development. Scripts
set this way are not executed by the browsers. To circumvent this a
React component is inserted that gets re-rendered with the page. On
re-render this element replaces all the script tags in such a way that
they run properly.

This code is set-up to do nothing in production the script tags are
pre-baked there and execute normally.

This commit also adds the development version of React to the pages so
that snippets relying on React don't break.

In a future commit both these things should be disableable using plugin
configuration. Additionally the version of React/ReactDOM added globally
should switch to production when the site is built for production mode.